Fixed signed and unsigned comparison
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@4061 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Samplers/GridSampler.h
@@ -91,7 +91,7 @@ class GridSampler : public SamplerMethod<MPTraits> {
       totalCell = totalCell* it->second;
     }
     //Generate all the points in the grid for the asked dimensions 
-    for(int iter=0; iter<totalCell; iter++){
+    for(size_t iter=0; iter<totalCell; iter++){
       map<size_t,size_t> coordinates;
       map<size_t,int> tempSize;
       int dimSize = 1;
